visualization-backend/requirements.txt
fastapi
uvicorn
torch
transformers
pillow
scikit-learn
matplotlib
opencv-python
